+++ MonoBank/BankManager.swift
@@ -9,38 +9,38 @@ import CoreData
 import Foundation
 import UIKit
 
-// Manages data
+/// Manages data
 
 class BankManager {
 	
 	static let shared = BankManager()
     
-	// Formats balances with the current currency symbol
+	/// Formats balances with the current currency symbol
 	let numberFormatter = NumberFormatter()
 	
-	// Currency symbol used by the app
+	/// Currency symbol used by the app
 	var currencySymbol = "$" {
 		didSet {
 			numberFormatter.currencySymbol = currencySymbol
 		}
 	}
 	
-	// Default balance when adding a player
-	var defaultBalance = 1500
+	/// Default balance when adding a player
+	var defaultBalance = Int32(1500)
 	
-	// Amount which the user can quickly add in the player menu
-	var quickAddAmount = 200
+	/// Amount which the user can quickly add in the player menu
+	var quickAddAmount = Int32(200)
 	
-	// All players of the current game
+	/// All players of the current game
 	var players = [Player]()
     
-    // All managedUser entities of the current game
+    /// All managedUser entities of the current game
     var managedUsers = [User]()
     
-    // All transactions of the current game
+    /// All transactions of the current game
     var transactions = [Transaction]()
     
-    // All managedReceipts entities of the current game
+    /// All managedReceipts entities of the current game
     var managedReceipts = [Receipt]()
     
     var index = Int16(0)
@@ -54,10 +54,10 @@ class BankManager {
 			self.currencySymbol = currencySymbol
 		}
 		if let defaultBalance = UserDefaults.standard.value(forKey: "defaultBalance") as? Int {
-			self.defaultBalance = defaultBalance
+			self.defaultBalance = Int32(defaultBalance)
 		}
         if let quickAddAmount = UserDefaults.standard.value(forKey: "quickAddAmount") as? Int {
-            self.quickAddAmount = quickAddAmount
+            self.quickAddAmount = Int32(quickAddAmount)
         }
         // Loading Users from core data
         if let appDelegate = UIApplication.shared.delegate as? AppDelegate {
@@ -72,13 +72,13 @@ class BankManager {
             let receipts = try managedContext.fetch(Receipt.fetchRequest()) as [Receipt]
             
             for user in users {
-                let player = Player(name: user.name!, balance: Int(user.balance), token: Token(rawValue: user.token!)!)
+                let player = Player(name: user.name!, balance: user.balance, token: Token(rawValue: user.token!)!)
                 players.append(player)
                 managedUsers.append(user)
             }
             
             for receipt in receipts {
-                let transaction = Transaction(amount: Int(receipt.amount), payee: receipt.payee!, payeeIndex: Int(receipt.payeeIndex), payer: receipt.payer!, payerIndex: Int(receipt.payerIndex))
+                let transaction = Transaction(amount: receipt.amount, payee: receipt.payee!, payeeIndex: Int(receipt.payeeIndex), payer: receipt.payer!, payerIndex: Int(receipt.payerIndex))
                 transactions.append(transaction)
                 managedReceipts.append(receipt)
             }
@@ -98,7 +98,7 @@ class BankManager {
 		numberFormatter.currencySymbol = currencySymbol
 	}
 	
-	//Saves the current settings of the BankManager
+	///Saves the current settings of the BankManager
 	func saveSettings() {
 		UserDefaults.standard.set(currencySymbol, forKey: "currencySymbol")
 		UserDefaults.standard.set(defaultBalance, forKey: "defaultBalance")
@@ -106,30 +106,30 @@ class BankManager {
 		UserDefaults.standard.set(soundsEnabled, forKey: "soundsEnabled")
 	}
     
-    //Saves the newly created Player to core data
+    ///Saves the newly created Player to core data
     func savePlayer(_ player: Player) {
         
         let user = User(entity: User.entity(), insertInto: CoreDataStack.managedContext)
         user.name = player.name
-        user.balance = Int16(player.balance)
+        user.balance = player.balance
         user.token = player.token.rawValue
         user.index += index
         index += 1
         
         managedUsers.append(user)
         
-        //In case if user moved collectionView cells before adding players
+        //In case if 'User' moved collectionView cells before adding players
         for i in (0..<managedUsers.count){
             managedUsers[i].index = Int16(i)
         }
         CoreDataStack.appDelegate.saveContext()
     }
     
-    //Saves the newly created Transaction to core data
+    ///Saves the newly created Transaction to core data
     func saveTransaction(_ transaction: Transaction){
         
         let receipt = Receipt(entity: Receipt.entity(), insertInto: CoreDataStack.managedContext)
-        receipt.amount = Int16(transaction.amount)
+        receipt.amount = transaction.amount
         receipt.payee = transaction.payee
         receipt.payeeIndex = Int16(transaction.payeeIndex)
         receipt.payer = transaction.payer
